+++ src/collective/tiles/collection/vocabularies.py
@@ -1,61 +1,46 @@
 # -*- coding: utf-8 -*-
 from collective.tiles.collection.interfaces import ICollectionTileRenderer
 from plone import api
-from plone.api.exc import InvalidParameterError
-from plone.app.customerize import registration
+from zope.component import getSiteManager
 from zope.globalrequest import getRequest
-from zope.i18n import translate
 from zope.interface import implementer
-from zope.publisher.interfaces.browser import IBrowserRequest
+from zope.interface import Interface
+from zope.interface import providedBy
 from zope.schema.interfaces import IVocabularyFactory
+from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
-from six.moves import filter
-from six.moves import map
 
 
 @implementer(IVocabularyFactory)
 class CollectionRenderersVocabulary(object):
     def __call__(self, context):
-        views = registration.getViews(IBrowserRequest)
-        renderers = list(filter(self.isCollectionRenderer, views))
-        lst = sorted(map(self.generateTerms, renderers), key=lambda k: k.title)
-
-        return SimpleVocabulary(lst)
-
-    def isCollectionRenderer(self, view):
         """
-        Filter only views that implements a certain interface.
-        Also checks if the use can access to that view.
-        An user can't access to the view for different reasons:
-        - invalid permissions
-        - invalid browserlayer (product with the view not installed)
+        Return a list of registered views for ICollectionTileRenderer
         """
-        if not ICollectionTileRenderer.implementedBy(view.factory):
-            return False
-
+        # copied from plone.api "get_view" method
+        sm = getSiteManager()
         portal = api.portal.get()
         request = getRequest()
-        # check if the view is registered for the current layer
-        for iface in view.required:
-            if not iface.providedBy(request):
-                return False
 
-        # try to access this view. If the use can't access the view,
-        # return False
-        try:
-            view = api.content.get_view(context=portal, name=view.name, request=request)
-            return view and True or False
-        except InvalidParameterError:
-            return False
+        available_views = sm.adapters.lookupAll(
+            required=(providedBy(portal), providedBy(request)), provided=Interface
+        )
+        renderers = []
+        for name, factory in available_views:
+            if not ICollectionTileRenderer.implementedBy(factory):
+                continue
+            renderers.append(self.generateTerms(name=name, factory=factory))
 
-    def generateTerms(self, view):
+        return SimpleVocabulary(sorted(renderers, key=lambda k: k.title))
+
+    def generateTerms(self, name, factory):
         """
         Return a SimpleTerm with id and translated view name
         """
-        factory = view.factory
-        name = view.name
         human_name = getattr(factory, "display_name", name)
-        return SimpleVocabulary.createTerm(name, name, api.portal.translate(human_name))
+        return SimpleTerm(
+            value=name, token=name, title=api.portal.translate(human_name)
+        )
 
 
 CollectionRenderersVocabularyFactory = CollectionRenderersVocabulary()